pyproject.toml
[project]
name = "leomed-bench"
version = "0.1.0"
description = "SLURM benchmark for 1 vs 4 GPU Swin training on ImageNet-1k."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jaxtyping>=0.3.9",
    "lightning>=2.6.1",
    "pyyaml>=6.0.3",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.4",
    "ty>=0.0.19",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/leomed_bench"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "I",
    "UP",
    "ANN",
    "NPY",
]
ignore = ["F722", "F821", "UP037", "UP040"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN"]

[tool.ty.environment]
python-version = "3.12"
